test: ExamRepository 테스트 추가
